Rename ArenaConfigBuilder to ArenaConfigLoader
