+++ src/RobotSimulator/ArenaConfigLoader.ts
